+++ jest.config.js
@@ -1,9 +1,7 @@
 module.exports = {
-    testEnvironment: "jsdom",
-    moduleNameMapper: {
-        "\\.(scss|sass|css)$": "identity-obj-proxy",
-    },
-    setupFilesAfterEnv: [
-        "<rootDir>/src/jest/setupTests.ts"
-    ]
+  testEnvironment: 'jsdom',
+  moduleNameMapper: {
+    '\\.(scss|sass|css)$': 'identity-obj-proxy',
+  },
+  setupFilesAfterEnv: ['<rootDir>/src/jest/setupTests.ts'],
 }