+++ src/components/Header/Header.tsx
@@ -1,13 +1,28 @@
-import React from 'react'
+import React, { useState, useRef } from 'react'
+import cx from 'classnames'
 
 import s from './Header.scss'
 
 import { useAppSelector } from '@msp/redux/hooks'
 import BellIcon from '@msp/components/icons/BellIcon'
 import UserIcon from '@msp/components/icons/UserIcon'
+import Logo from '@msp/components/common/Logo'
+import useOnClickOutside from '@msp/hooks/useOnClickOutside'
+import UserMenu from './components/UserMenu'
+import { useLocationChange } from '@msp/hooks/useLocationChange'
+
+const CLICK_OUTSIDE_IGNORE_CLASSNAME = 'ignoreUserMenu'
 
 const Header: React.FC = () => {
   const { userInfo } = useAppSelector((state) => state.user)
+  const [isMenuShown, setMenuShown] = useState(false)
+  // close the menu on navigation
+  useLocationChange(() => {
+    onCloseMenu()
+  })
+  const menuRef = useRef(null)
+  const onCloseMenu = () => setMenuShown(false)
+  useOnClickOutside(menuRef, onCloseMenu, CLICK_OUTSIDE_IGNORE_CLASSNAME)
 
   return (
     <div className={s.container}>
@@ -21,10 +36,19 @@ const Header: React.FC = () => {
         <div className={s.notificationsContainer}>
           <BellIcon />
         </div>
-        <div className={s.userDetails}>
-          <img className={s.companyLogo} src={userInfo!.logo} />
-          <span className={s.name}>{userInfo!.name}</span>
-          <UserIcon />
+        <div className={s.userDetailsContainer}>
+          <div
+            className={cx(s.userDetails, CLICK_OUTSIDE_IGNORE_CLASSNAME)}
+            onClick={() => setMenuShown(!isMenuShown)}>
+            <Logo link={userInfo!.logo!} />
+            <span className={s.name}>{userInfo!.name}</span>
+            <UserIcon />
+          </div>
+          {isMenuShown ? (
+            <div ref={menuRef}>
+              <UserMenu userInfo={userInfo!} />
+            </div>
+          ) : null}
         </div>
       </div>
     </div>