+++ src/components/Overview/types.d.ts
@@ -1,0 +1,4 @@
+export type linkType = {
+  linkTitle?: string
+  linkAddress: string
+}