+++ src/routes/main.ts
@@ -1,3 +1,5 @@
+import React from 'react'
+import CallBack from '@msp/components/CallBack'
 import Overview from '@msp/components/Overview'
 import Invoices from '@msp/components/Billing/Invoices'
 
@@ -9,6 +11,7 @@ interface RouteProps {
 const routes: RouteProps[] = [
   { path: '/', element: Overview },
   { path: '/invoices', element: Invoices },
+  { path: '/callback', element: CallBack },
 ]
 
 export default routes