+++ CHANGELOG.md
@@ -2,19 +2,64 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
-### 0.1.3 (2022-10-10)
+### [0.1.10](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/compare/v0.1.9...v0.1.10) (2022-10-12)
 
 
 ### ✨ Features
 
-* **asdas:** sadas ([93195a1](https://github.com/ivankraev/msb-ui/commit/93195a1a723a47dc348fab5f2615f85e5ba8ec7e))
+* added code challenge generator ([747097c](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/commit/747097cd0f402f3aaf8072589bfb8d3508331923))
+* added envs example ([354db5e](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/commit/354db5e69cf5439eba7731ceaff8efe942b0b40f))
+* added eslint unused imports rule ([30bac15](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/commit/30bac159e8f99b19d4d393e018af741c918782fa))
+* okta user sso ([3bc9dd8](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/commit/3bc9dd8e930a3b30dd331230274a37eaa6bc10f7))
 
-### [0.1.2](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/compare/v0.1.1...v0.1.2) (2022-10-09)
+
+### 🚚 Chores
+
+* added axios, webpackenv ([049aa66](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/commit/049aa66e047443135c5fed78076c04fbe75c243f))
+* Merge branch 'main' into feature/okta ([fb30753](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/commit/fb30753cb30805d9fff3e52287c7c1b63291de3d))
+
+
+### 🐛 Bug Fixes
+
+* after merge fix ([9b56dd0](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/commit/9b56dd0a66693038bd687d84ef842df62c49bd77))
+* after merge fix ([444e18d](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/commit/444e18d2cc451fe049ef73bb2d5b8fba131e0837))
+* formatted with prettier ([5b57ad7](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/commit/5b57ad7d5b147e85c0fc010084e0aea07443619b))
+
+
+### ♻️ Code Refactoring
+
+* **custom components:** resolved conflicts ([4056ff9](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/commit/4056ff9333274895343085b9d616a5ae6a6832bf))
+
+### [0.1.9](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/compare/v0.1.7...v0.1.9) (2022-10-12)
+
+
+### ✅ Testing
+
+* **custom copmponents:** added tests to all the components ([7e1ca84](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/commit/7e1ca847a2f6147580ea0e43e80c2c460ec69591))
+
+### [0.1.8](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/compare/v0.1.1...v0.1.8) (2022-10-12)
 
 
 ### ✨ Features
 
-* **invoices:** added Invoices Page and created Custom table ([4546321](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/commit/45463212d916fbb32840971eb012c652c8d482a3))
+* added non-active ([b3075fb](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/commit/b3075fbd50d4427d96f41568d8a467ce5f713fcb))
+* **jest unit test:** added jest config and basic Overview page tests ([8ca2d00](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/commit/8ca2d00a31dbd98317496fb48dbaf3bf235b86ac))
+* overview page ([de88f25](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/commit/de88f25c47ce06a5759b2838047d3815f4a7f95d))
+
+
+### 🐛 Bug Fixes
+
+* fixed BigCard component unit tests ([4d86176](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/commit/4d861762f3599dc06b1d7f0741a17cb61cb3c59c))
+* fixed h1 styles ([d9b660b](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/commit/d9b660b3aca6fd9e67365d1ee140e44ddf407829))
+* fixed styles with vars ([ac9e4ca](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/commit/ac9e4ca0b10d7515b5181ab259619640574aa768))
+* fixed types file ([b08ba38](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/commit/b08ba385e63f4336551eb005b375089925f40d14))
+* **smallcard class:** added backup class for smallCard notactive component ([3a26572](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/commit/3a2657212762b688ac1677ad38647bf0113baf66))
+* ts fix ([d5b3434](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/commit/d5b34344bd88b5fec9e90ac91d6fcb74ebf4333e))
+
+
+### ✅ Testing
+
+* **custom components:** added tests to all the components ([439444c](https://mstfs.softserveinc.com/tfs/Cisco/MSP/_git/UI/commit/439444cdc71ab407e10a496500f5463f21968d53))
 
 ### 0.1.1 (2022-10-07)